# docs/swagger.yaml
basePath: /
definitions:
...
  dto.LoginResponse:
    properties:
      access_token:
        type: string
      expires_in:
        type: number
      token_type:
        type: string
    type: object
  dto.MessageResponse:
    properties:
      message:
        type: string
    type: object
  dto.UserLoginRequest:
    properties:
      login:
        type: string
      password:
        type: string
    required:
    - login
    - password
    type: object
  dto.UserRegisterRequest:
    properties:
      login:
        maxLength: 25
        minLength: 3
        type: string
      password:
        minLength: 6
        type: string
    required:
    - login
    - password
    type: object
  dto.UserResponse:
    properties:
      id:
        type: string
      login:
        type: string
      role:
        $ref: '#/definitions/role.Role'
    type: object
  dto.UserUpdateRequest:
    properties:
      login:
        maxLength: 25
        minLength: 3
        type: string
      password:
        maxLength: 100
        minLength: 6
        type: string
    type: object
  role.Role:
    enum:
    - 0
    - 1
    - 2
    type: integer
    x-enum-comments:
      Admin: "2"
      Guest: "0"
      User: "1"
    x-enum-descriptions:
    - "0"
    - "1"
    - "2"
    x-enum-varnames:
    - Guest
    - User
    - Admin
host: localhost:8080
info:
  contact:
    email: kovalyovea@student.bmstu.ru
    name: API Support
    url: https://github.com/nylox23/DIA
  description: Calculates the volume of gas emitted in a reaction
  license:
    name: AS IS (NO WARRANTY)
  title: Carb_Calc
  version: "1.0"
paths:
...
  /api/carbonate-acids/{id}:
    delete:
      description: Удаляет кислоту из текущей заявки-черновика пользователя
      security:
        - Bearer: []
      parameters:
      - description: ID кислоты
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.MessageResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Удалить кислоту из заявки
      tags:
      - carbonate-acids
    put:
      consumes:
      - application/json
      description: Обновляет массу кислоты в текущей заявке-черновике пользователя
      security:
        - Bearer: []
      parameters:
      - description: ID кислоты
        in: path
        name: id
        required: true
        type: integer
      - description: Масса кислоты
        in: body
        name: request
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.MessageResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Обновить количество кислоты
      tags:
      - carbonate-acids
...
  /api/users/login:
    post:
      consumes:
      - application/json
      description: Аутентифицирует пользователя с помощью логина и пароля, возвращает
        JWT токен
      parameters:
      - description: Данные для входа в систему
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UserLoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.LoginResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Aутентификация пользователя
      tags:
      - users
  /api/users/logout:
    post:
      consumes:
      - application/json
      description: Добавляет JWT токен в блеклист, деактивируя его
      security:
        - Bearer: []
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.MessageResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Разлогинить пользователя
      tags:
      - users
  /api/users/profile:
    get:
      description: Возвращает информацию из профиля пользователя
      security:
        - Bearer: []
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.UserResponse'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Получить данные пользователя
      tags:
      - users
    put:
      consumes:
      - application/json
      description: Обновляет информацию в профиле пользователя (логин и/или пароль)
      security:
        - Bearer: []
      parameters:
      - description: Данные для обновления пользователя
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UserUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.UserResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Обновить данные пользователя
      tags:
      - users
  /api/users/register:
    post:
      consumes:
      - application/json
      description: Создает нового пользователя с предоставленными логином и паролем
      parameters:
      - description: Данные для регистрации пользователя
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UserRegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.UserResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Зарегистрировать пользователя
      tags:
      - users
schemes:
- http
swagger: "2.0"
securityDefinitions:
  Bearer:
    type: apiKey
    name: Authorization
    in: header